调整目录结构

+++ src/main/java/com/johnny/althorithm/bits/algo/datastruct/bst/CRUDSolution.java
@@ -58,7 +58,7 @@ public class CRUDSolution {
     }
 
 
-    public TreeNode remove(TreeNode root, int target) {
+    public TreeNode remove(TreeNode root, long target) {
         if (null == root) {
             return null;
         }
